lv.8 완료

+++ NBC_PhoneBook/NBC_PhoneBook/source/Models/PhoneBookManager.swift
@@ -5,28 +5,28 @@
 //  Created by t2023-m0072 on 12/8/24.
 //
 
-import Foundation
+import UIKit
 
 import CoreData
 
 /// CRUD
 class PhoneBookManager {
     
+    static let shared = PhoneBookManager()
+    
     private let container: NSPersistentContainer
+    private let entity: NSEntityDescription?
     
-    private lazy var entity = NSEntityDescription.entity(forEntityName: PokePhoneNumberData.className,
-                                                         in: self.container.viewContext)
-    
-    init(container: NSPersistentContainer) {
+    private init() {
+        let appDelegate = UIApplication.shared.delegate as! AppDelegate
+        let container = appDelegate.persistentContainer
         self.container = container
+        self.entity = NSEntityDescription.entity(forEntityName: PhoneNumberData.className,
+                                                 in: self.container.viewContext)
     }
     
-    func creat(_ pokePhoneNumber: PokePhoneNumber) {
-        guard let entity = self.entity,
-              let pokePhoneNumberData = NSManagedObject(entity: entity,
-                                                        insertInto: self.container.viewContext) as? PokePhoneNumberData else { return }
-        
-        pokePhoneNumberData.setPokePhoneNumber(pokePhoneNumber)
+    func creat(_ phoneNumber: PhoneNumber) {
+        make(by: phoneNumber)
         
         do {
             try save()
@@ -36,52 +36,64 @@ class PhoneBookManager {
         
     }
     
-    func read() -> [PokePhoneNumber]? {
-        var pokePhoneNumbers: [PokePhoneNumber] = []
+    func read() -> [PhoneNumber]? {
+        guard let phoneNumberDatas = phoneNumberDatas() else { return nil }
+        let phoneNumbers = phoneNumberDatas.compactMap { $0.convertTo() }
         
-        do {
-            let pokePhoneNumberDatas = try self.container.viewContext.fetch(PokePhoneNumberData.fetchRequest())
-            pokePhoneNumbers = pokePhoneNumberDatas.compactMap { $0.pokePhoneNumber() }
-            
-        } catch let error {
-            print(error)
-        }
-        
-        return !pokePhoneNumbers.isEmpty ? pokePhoneNumbers : nil
+        return phoneNumbers
     }
     
-    func update(_ pokePhoneNumber: PokePhoneNumber) {
+    func update(_ phoneNumber: PhoneNumber) {
+        guard let phoneNumberDatas = phoneNumberDatas() else { return }
+        phoneNumberDatas.forEach { $0.update(phoneNumber) }
+        
         do {
-            let pokePhoneNumberDatas = try self.container.viewContext.fetch(PokePhoneNumberData.fetchRequest())
-            pokePhoneNumberDatas.forEach {
-                $0.update(pokePhoneNumber)
-            }
-            
             try save()
-            
         } catch let error {
-            print(error)
+            print("update: error - \(error.localizedDescription)")
         }
     }
     
-    func delete(_ pokePhoneNumber: PokePhoneNumber) {
+    func delete(_ phoneNumber: PhoneNumber) {
+        guard let phoneNumberDatas = phoneNumberDatas() else { return }
+        
+        phoneNumberDatas.forEach {
+            if $0.isSame(phoneNumber) { container.viewContext.delete($0) }
+        }
         do {
-            let pokePhoneNumberDatas = try self.container.viewContext.fetch(PokePhoneNumberData.fetchRequest())
-            pokePhoneNumberDatas.forEach {
-                if $0.isSame(pokePhoneNumber) {
-                    container.viewContext.delete($0)
-                }
-            }
-            
             try save()
-            
         } catch let error {
-            print(error)
+            print("delete: error - \(error.localizedDescription)")
+        }
+    }
+    
+    func reset() {
+        guard let phoneNumberDatas = phoneNumberDatas() else { return }
+        
+        phoneNumberDatas.forEach {
+            container.viewContext.delete($0)
+        }
+        do {
+            try save()
+        } catch let error {
+            print("delete: error - \(error.localizedDescription)")
         }
     }
     
     
     private func save() throws {
         try self.container.viewContext.save()
+    }
+    
+    private func make(by phoneNumber: PhoneNumber) {
+        guard let entity else { return  }
+        let object = NSManagedObject(entity: entity, insertInto: self.container.viewContext)
+        
+        guard let phoneNumberData = object as? PhoneNumberData else { return  }
+        phoneNumberData.setPhoneNumber(phoneNumber)
+    }
+    
+    private func phoneNumberDatas() -> [PhoneNumberData]? {
+        return try? self.container.viewContext.fetch(PhoneNumberData.fetchRequest())
     }
 }
